+++ package.json
@@ -10,8 +10,9 @@
   "main": "dist/ko-component-router.js",
   "scripts": {
     "build": "scripts/build.js",
+    "watch": "scripts/watch.js",
+    "test": "scripts/test.js",
     "lint": "eslint --max-warnings=0 *.js",
-    "test": "karma start",
     "prepublish": "npm run build"
   },
   "keywords": [
@@ -32,21 +33,30 @@
   },
   "devDependencies": {
     "babel-loader": "^6.2.0",
+    "babel-plugin-transform-async-functions": "^6.8.0",
     "babel-plugin-transform-async-generator-functions": "^6.17.0",
+    "babel-plugin-transform-es2015-for-of": "^6.18.0",
     "babel-plugin-transform-object-rest-spread": "^6.19.0",
+    "babel-plugin-transform-regenerator": "^6.20.0",
     "babel-preset-es2015": "^6.1.18",
+    "babel-preset-es2015-rollup": "^1.2.0",
     "babel-preset-es2017": "^6.16.0",
+    "babel-runtime": "^6.20.0",
     "chalk": "^1.1.3",
+    "chokidar": "^1.6.1",
+    "concurrently": "^3.1.0",
     "coveralls": "^2.11.9",
     "eslint": "^3.0.1",
     "eslint-config-profiscience": "^1.0.1",
     "filesize": "^3.3.0",
     "isparta-loader": "^2.0.0",
+    "istanbul-instrumenter-loader": "^1.1.0",
     "jquery": "^3.1.1",
     "karma": "^1.1.0",
     "karma-chrome-launcher": "^2.0.0",
     "karma-coverage": "^1.0.0",
     "karma-firefox-launcher": "^1.0.0",
+    "karma-rollup-preprocessor": "^3.0.1",
     "karma-tap": "^3.0.1",
     "karma-webpack": "^1.7.0",
     "knockout": "^3.3.0",
@@ -56,6 +66,7 @@
     "regenerator-runtime": "^0.10.0",
     "rimraf": "^2.5.4",
     "rollup": "^0.36.4",
+    "rollup-plugin-babel": "^2.6.1",
     "rollup-plugin-buble": "^0.14.0",
     "rollup-plugin-commonjs": "^5.0.5",
     "rollup-plugin-node-resolve": "^2.0.0",
